# Implement `Vec::from_elem` specialization for all `Copy` types