If the input element is zero, `Vec::from_elem` can just invoke
`calloc` for any `Copy` type.

If the input is non-zero, but its size is 1, it can allocate and then
`memset` the buffer.