@BurntSushi 2017-04-19
cc @rust-lang/libs