@ranma42 2017-04-20
@alexcrichton Could you explain a little more what is the issue with padding? AFAICT in the worst case the input object might have some padding initialized to a non-zero value, in which case it would not benefit from `calloc`.

From what I understood about coherence rules and specialization, it would be impossible to customize the implementation for several types for which this would be useful, such as `Option`, tuples and small arrays.