@aturon 2017-04-21
Do we have evidence of this making a performance difference?