@alexcrichton 2017-04-21
@ranma42 ah yes sorry, I mean what you're thinking. A "zero" type like `(0u8, 0u16)` is not guaranteed to receive this optimization, but it may sporadically do so. That means that you could in development see the optimization, start to rely on it, and then it doesn't work on production mysteriously.