@ranma42 2017-04-21
@alexcrichton yes, that can be a problem, but on the bright side it would not affect the correctness of the program: the "issue" would be that in some conditions the program run faster than you expect, instead of always being slow. Moreover, the same holds true for most optimisations (especially when combined). In some specific contexts (avoiding timing attacks in cryptography, meeting guaranteed deadlines in realtime) predictable timings are more desirable than fast programs, but I believe this is not the general case nor something that LLVM has strong support for.

@aturon I rerun the benchmark from https://github.com/rust-lang/rust/pull/40409#issuecomment-285704013 adapted as follows:
```rust
#![feature(test)]

extern crate test;

use test::Bencher;

#[derive(Clone)]
struct Clonable<T>(T);

#[bench]
fn bench_data(b: &mut Bencher) {
    b.iter(|| vec![0u8; 1024 * 1024]);
}

#[bench]
fn bench_clone(b: &mut Bencher) {
    b.iter(|| vec![Clonable(0u8); 1024 * 1024]);
}

#[bench]
fn bench_copy(b: &mut Bencher) {
    b.iter(|| vec![(0u8,); 1024 * 1024]);
}

#[bench]
fn bench_nullptr(b: &mut Bencher) {
    b.iter(|| vec![std::ptr::null::<u8>(); 1024 * 1024]);
}

#[bench]
fn bench_none_ref(b: &mut Bencher) {
    b.iter(|| vec![None::<&u8>; 1024 * 1024]);
}

#[bench]
fn bench_padded(b: &mut Bencher) {
    b.iter(|| vec![(0u8,0u16); 1024 * 1024]);
}
```

shows these results:

Before:
```
test bench_clone    ... bench:     249,831 ns/iter (+/- 165)
test bench_copy     ... bench:     249,943 ns/iter (+/- 559)
test bench_data     ... bench:         956 ns/iter (+/- 5)
test bench_none_ref ... bench:     625,907 ns/iter (+/- 872)
test bench_nullptr  ... bench:     625,807 ns/iter (+/- 821)
test bench_padded   ... bench:     897,815 ns/iter (+/- 1,117)
```

After
```
test bench_clone    ... bench:     249,393 ns/iter (+/- 193)
test bench_copy     ... bench:         941 ns/iter (+/- 7)
test bench_data     ... bench:         940 ns/iter (+/- 6)
test bench_none_ref ... bench:         721 ns/iter (+/- 2)
test bench_nullptr  ... bench:         719 ns/iter (+/- 1)
test bench_padded   ... bench:         749 ns/iter (+/- 4)
```