@sfackler 2017-04-21
Doesn't reading padding bytes produce a poison value?