@ranma42 2017-04-21
@sfackler yes, that is what @alexcrichton mentioned: reading padding bytes can result in an `undef` as a result for `is_zero`, which eventually leaves the compiler with the freedom to choose whether using the `calloc` path or the slow one.