@sfackler 2017-04-21
Is that the extent of it? Are you sure it's not going to poison other future computation?