@rkruppe 2017-04-21
If padding bytes are merely `undef`, current semantics are that the compiler can pick an arbitrary value for them, and possibly a different one every time they're read. This is very problematic in most contexts (e.g., if you use it as an index, you might see different values in the bounds check vs. in the indexing, breaking memory safety) but in this case I can't see a way for that to break (even if you get a different value during the actual copy, you're writing to padding bytes so it shouldn't matter). That doesn't necessarily mean it's fine, though, and it makes me very uneasy to rely on such tricky minutia!

However, `undef` may be slowly phased out in favor of a revised `poison` value, see http://www.cs.utah.edu/~regehr/papers/undef-pldi17.pdf &mdash; the gist is that reading uninitialized memory (which padding is, probably) would indeed give you a poison value, and poison propagates through all computations. This means `is_zero`'s result would be poison if there are any padding bytes, and branching on poison is UB.

In the same breath, a `freeze` instruction is added, which fixes an arbitrary value for a poison value (like current `undef`), but it's far from clear to me by what rule rustc should be inserting `freeze` operations into `is_zero` (and if it does, whether that would impact optimizations).

And of course, all this is only as far as LLVM is defined, the unsafe code guidelines probably have or will have an opinion on reading padding bytes, too.

tl;dr Reading padding bytes is a minefield.