@alexcrichton 2017-04-21
@rkruppe so to clarify, you think that this PR leads to undefined behavior? If so that seems quite bad!