@rkruppe 2017-04-21
I *don't think* it currently leads to UB as far as LLVM is concerned. However,

1. I'm not 100% sure.
2. LLVM is a moving target.
3. The unsafe code guidelines may differ.