@ranma42 2017-04-21
@rkruppe I think the paper proposes a nice direction to improve the consistency of handling `undef` values across different passes in LLVM. AFAICT the same approach they propose for bitfields could apply just as well to any composite data structure, even if it includes padding.

@alexcrichton yes, as per https://github.com/rust-lang/rust/pull/41335#issuecomment-296126852, that is expected. My understanding of LLVM semantics is that when a jump depends on an `undef` right now LLVM is required to (non-deterministically) choose one of the branches. It looks like this might change in the future, but this would most likely be a breaking change (the removal of `undef` in LLVM would be breaking as well).

The safest route is certainly to close without merging. It is a pity that the handling of `Copy` types is not unified and that several type which could benefit from this are not covered, but it looks like Rust cannot express the low-level check `is_zero` in a satisfiable way.

Part of the results of this PR could be achieved by explicitly implementing the trait for `*{const,mut} T` and `Option<&T>` (the two types which I would expect to gain the most from the optimization, as they can be used to create an array of null pointers).