@rkruppe 2017-04-22
> AFAICT the same approach they propose for bitfields could apply just as well to any composite data structure, even if it includes padding.

Yes, but we most likely wouldn't want to do it for padding, at least not everywhere. It bloats IR, possibly inhibits optimizations, and is entirely pointless for all code except code that tries to be too smart for its own good about padding bytes.